requirements_patch.txt
# Add these to your requirements.txt if missing
google-generativeai
Pillow
streamlit-drawable-canvas
gTTS
SpeechRecognition
streamlit-mic-recorder
